switch icons to react-icons, add social icons, adjust section styling a bit
