+++ src/components/Home/AboutMe.tsx
@@ -9,7 +9,7 @@ const AboutMe: React.FC = () => {
   return (
     <section id="about" className="bg-[#f6e3ceaf] border-y border-stone-300">
       <div className="grid lg:container lg:mx-auto px-6 pb-6 text-center">
-        <Heading className="mb-10">About Michelle</Heading>
+        <Heading className="mb-6">About Michelle</Heading>
         <div className="grid md:grid-cols-2 gap-6">
           <div className="hidden md:block ">
             <Image
