+++ src/components/common/Header.tsx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react';
 import Link from 'next/link';
 import { Popover, Transition } from '@headlessui/react';
-import { MenuIcon, XIcon } from '@heroicons/react/outline';
+import { HiMenuAlt2, HiX } from 'react-icons/hi';
 
 const Header: React.FC = () => {
   const handleClick = (e: React.MouseEvent<HTMLAnchorElement>) => {
@@ -30,7 +30,7 @@ const Header: React.FC = () => {
           <div className="-mr-2 -my-2 md:hidden">
             <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
               <span className="sr-only">Open menu</span>
-              <MenuIcon className="h-6 w-6" aria-hidden="true" />
+              <HiMenuAlt2 className="h-6 w-6" aria-hidden="true" />
             </Popover.Button>
           </div>
           <nav className="hidden md:flex space-x-12">
@@ -86,7 +86,7 @@ const Header: React.FC = () => {
                 <div className="-mr-2">
                   <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
                     <span className="sr-only">Close menu</span>
-                    <XIcon className="h-6 w-6" aria-hidden="true" />
+                    <HiX className="h-6 w-6" aria-hidden="true" />
                   </Popover.Button>
                 </div>
               </div>
